refactor: 어노테이션 디렉토리 이동
