+++ src/main/java/com/pitchain/dto/res/ReplyCommentRes.java
@@ -1,6 +1,6 @@
 package com.pitchain.dto.res;
 
-import com.pitchain.common.apiPayload.annotation.S3Url;
+import com.pitchain.common.annotation.S3Url;
 import com.pitchain.entity.Comment;
 import com.pitchain.entity.Member;
 import jakarta.validation.constraints.NotEmpty;
